Restructure Audio Loudness details in DM XML
We decided this DM isn't intended to be a report of Loudness tests that
were carried out on the file - it's just an indication of the Loudness
Specification that the programme-maker was intending for the content to
comply with.

This also removes the "Intentionally_Quiet" elements as that scenario
can be covered by the new structure.

Resolves AMWA-TV/AS-11_Overview#5