Refactor MCA Labeling Framework Blocks
This permits the use of this Block for Soundfields comprised of
multiple, single-channel Sound Tracks as well as those comprised of a
single, multi-channel Sound Track.

This commit also tidies up the "Audio Layout" Blocks, to clarify how the
"Audio Layout Mode"s work.

The preamble to the tables of details of MCA Labeling labels is also
improved.